Heavily refactored the codebase to accomodate the spell system
