--- src/main/java/com/ben/wandwars/spells/components/blockHit/NoBlockHitEffect.java
@@ -1,13 +1,0 @@
-package com.ben.wandwars.spells.components.blockHit;
-
-import com.ben.wandwars.Main;
-import org.bukkit.Location;
-
-import java.util.UUID;
-
-public class NoBlockHitEffect extends BlockHitEffect {
-    @Override
-    public void hit(Location location, Main main, UUID caster) {
-
-    }
-}
