+++ src/main/java/com/ben/wandwars/wands/Wands.java
@@ -1,7 +1,6 @@
 package com.ben.wandwars.wands;
 
 
-import com.ben.wandwars.wands.items.ExplosionWand;
 import com.ben.wandwars.wands.wandInterfaces.Wand;
 import org.bukkit.inventory.ItemStack;
 
@@ -17,8 +16,6 @@ public class Wands {
     List<Wand> WIPWands = new ArrayList<>();
     private Wands() {
         //add the wands to the list
-        wands.add(new ExplosionWand());
-        WIPWands.add(new ExplosionWand());
     }
